app/requirements.txt
django>=3.2
 djangorestframework>=3.12
 gunicorn>=20.0
 mysqlclient>=2.2.0
